First Draft Animatic
This is an animatic of the storyboards as of 1/5/13. Timing is rough,
and requires adjustments in several areas, although the boards seem to
read well at first glance. Apart from timing, some shots require
tweaking or mild restaging to read properly. Muting is recommended, as
the exporter did not handle the blank soundtrack properly.