Final Version of Animatic: Requesting Sign-Off
Requesting sign-off on the final version of Animatic (without sound).
This edition contains the revised ending provided by Christian
Williams, and has adjusted the timing of the ending to fit the new
boards. Once approved, we are good to move on to sound and color comps.

If additional tweaks are required, I am available from 2PM to 3PM on
Wednesday this week to fix and re-export. Any fixes that cannot be
repaired then will either have to wait until after the 17th, or be
assigned to another team member.